+++ cli/testrun/testrun.go
@@ -12,8 +12,6 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/neilotoole/sq/libsq/core/ioz"
-
 	"github.com/stretchr/testify/require"
 
 	"github.com/neilotoole/slogt"
@@ -22,6 +20,7 @@ import (
 	"github.com/neilotoole/sq/cli/config"
 	"github.com/neilotoole/sq/cli/config/yamlstore"
 	"github.com/neilotoole/sq/cli/run"
+	"github.com/neilotoole/sq/libsq/core/ioz"
 	"github.com/neilotoole/sq/libsq/core/lg"
 	"github.com/neilotoole/sq/libsq/core/options"
 	"github.com/neilotoole/sq/libsq/source"