+++ libsq/driver/record.go
@@ -11,6 +11,7 @@ import (
 	"text/template"
 	"time"
 
+	"github.com/shopspring/decimal"
 	"go.uber.org/atomic"
 
 	"github.com/neilotoole/sq/libsq/core/errz"
@@ -117,6 +118,10 @@ func (x *StmtExecer) Close() error {
 // elements of rec listed in skipped.
 //
 // REVISIT: Do we need the skip mechanism at all?
+// REVISIT: This function implementation is an unholy mess. Much of it
+// dates from sq's earliest days, and it's been hacked on ever since.
+// Several of the code paths can probably never be reached.
+// It should be completely rewritten.
 //
 //nolint:funlen,gocognit,gocyclo,cyclop
 func NewRecordFromScanRow(meta record.Meta, row []any, skip []int) (rec record.Record, skipped []int) {
@@ -268,6 +273,15 @@ func NewRecordFromScanRow(meta record.Meta, row []any, skip []int) (rec record.R
 				rec[i] = nil
 			}
 
+		case *decimal.Decimal:
+			rec[i] = *col
+
+		case *decimal.NullDecimal:
+			if col.Valid {
+				rec[i] = col.Decimal
+			} else {
+				rec[i] = nil
+			}
 		case *time.Time:
 			rec[i] = *col
 
@@ -291,25 +305,6 @@ func NewRecordFromScanRow(meta record.Meta, row []any, skip []int) (rec record.R
 			rec[i] = int64(*col)
 		case *float32:
 			rec[i] = float64(*col)
-		}
-
-		if rec[i] != nil && meta[i].Kind() == kind.Decimal {
-			// Drivers use varying types for numeric/money/decimal.
-			// We want to standardize on string.
-			switch col := rec[i].(type) {
-			case *string:
-				// Do nothing, it's already string
-
-			case *[]byte:
-				rec[i] = string(*col)
-
-			case *float64:
-				rec[i] = stringz.FormatFloat(*col)
-
-			default:
-				// Shouldn't happen
-				rec[i] = fmt.Sprintf("%v", col)
-			}
 		}
 	}